Abandon project to absorb Route into Dispatcher.
The purpose of the project is to allow the rendering tree to be an AbstractFutureKeyedTree by passing the dependency list to `Dispatcher`s instead of having a separate Route class manage them. As it stood, Dispatcher was [and is being reverted to] a wrapper of a FastRoute\Dispatcher. The subtree relationship, starting at Dispatcher, used to be: [Dispatcher-{eventually}->]Route { path, resolver }->Map<dependency_key, Dispatcher>. it is fundamentally not possible for Dispatcher alone to maintain this subtree relationship, because the mediator of paths must be separate from the mediator of dependencies since they exist on different levels of resolution. The Dispatcher object needs to point paths to some List-like object maintaining a list of (dependency table, resolver) tuples. The dependency table points dependency keys to some object that maps paths. A single object can't fulfill both responsibilities elegantly — it would just be a union of the existing Route and Dispatcher objects that does one thing in one context and the other in another (e.g. as the value for a dependency key vs. the value for a path).

Furthermore, I've realized that Dispatcher shouldn't be a Tree at all — the `FastRoute\Dispatcher` object contains the subtree at every level, for which Dispatcher is just a wrapper. If `FastRoute\Dispatcher` were more strongly-typed, somewhere a structure would have to obey a tree structure.

This reverts commit c6c4cc7d5f168039a934bacd47a577bca52b12dc.