Reversing variance on key types (Tk for containers, Tx for trees) of classes either composing or inheriting from KeyedContainerWrapper from - to +.
\#\# Rationale
The most recent line of thought was to render Tk on all `KeyedContainerWrapper` descendants (Tx on `Tree`-ish classes) contravariant to accomodate a `get(Tk): Tv` method, which could more naturally be used by route resolvers. However, ` KeyedIterable`, which does implement such a method, is not variant not from a lack of foresight about the contravarian t possibility, but because `foreach`, which exposes the `Tk` key type, requires a covariantly-valid type. `KeyedContainer` can also be iterated using a keyed `foreach`, so it must also be valid covariantly on Tk.