Try to make a base type to the circular Comparable
Careless type errors aside, the types are still not sound, but this was
the closest attempt to making a ceiling type for `ComparableView`'s
`TCompatable` for State[Factoryie]s and Dispatcher, which don't care
about the specific ComparableView. I've since given up and just made a
`View` that disregards the comparison entirely.

The code typechecks on 3.19.1 and I thought it was interesting because
it looks like it may be okay: TComparable is swapped from a class
generic to a method generic with the same constraint between
`ComparableRoute` and `BaseRoute` respectively, but this actually isn't
sound. This sort of overriding is only safe if the parameter is in both
cases in a covariant position, because the upcast agrees with the
direction of the cast (for a generic method with an upper-bound, which
is probably the most natural case to think about, the generic is pushed
up to its upper bound. If this bound is consistent between both classes
then the bound :> the original type so the cast is valid covariantly.
It's a bit mind-bending to think about a lower-bound to a method generic
for me right now, but I think the covariant requirement is generally
true).

Weird that again two what I thought to be totally dispirate issues have
combined to a common problem:
[#8383](https://github.com/facebook/hhvm/issues/7383) and
[#7828](https://github.com/facebook/hhvm/issues/7828).