Backup before splitting Route implementations
The code is currently on track to implement `Route`, `Dispatcher` and
the like by stashing the views in `Route` and letting `Route` perform
the comparisons. Furthermore, `Route` produces an `XHPRoot` but is not
itself an XHP element. For expressive strength and the performance of
the comparison scheme, I am also going to experiment with `Route` as an
XHP element like `<:route:GET ... />`. I'm hoping to move the
responsibility of stashing and comparing to the calling scope, so that
the Dispatcher-Route tree can be as stateless as possible. This removes
the need for some awkward cloning which would be necessary to separate
the stashed views between different requests.