Saving progress on purging generics and flattening classes
What looked initially like a mild inconvenience has revealed that this
entire approach of managing the return types from resolvers is flawed.
Implementing `ViewTree` with unconstrained generics, I realized that it
was impossible to type without an unbounded number of generic terms.
This is because I was trying to assert that the current node's type
might be entirely disjoint from descendants, requiring more than one
type parameter. However, I also needed to break identity between the two
generics as I parameterized them for the children, so at every level I
needed an extra parameter in the parent than the child even though they
were the same class, which is obviously impossible.

Actually an orthogonal issue but arising when thinking of a solution to
this generic explosion, when aggregating all of these return types into
collections (like the dispatchers' route maps, or the views' subtree
map) the return type from the resolver (originally `TComparable`) would
be cast to the strongest common ancestor of _all_ resolvers, rather than
of any given resolver. In hindsight, this is unsurprising. I need as
many types as I have resolvers if I want to maintain their independence.
The core issue is that I am stashing `TComparable` in class properties
to be used between method calls on that class, and any state in classes
that is generic must be parameterized explicitly at the `class`
statement.

However, I think we might be able to salvage this behavior, by confining
the `TComparable` type within functions, and relying on type inference
to generate the types we need anonymously and opaquely. Then, these
functions would expose the less colorful types the parent scope can work
with (so the parent scope is no longer responsible for comparison) like
`\Stringish`, booleans, etc. I am pretty certain generators are the only
possible solution, so the next few commits will explore this solution.