Save progress on generator approach
I wanted to send "actions" into the generator to prompt it to re-render
the route and stage it (STAGE) or propagate an emission down the tree
(COMMIT). However, this requires that the tree partially split at every
render via clones and property setting and a lot of nonsense, so instead
I'm revisiting a tree-type return like the last commit, except without
generic explosion (the Generator, soon to be iterator, isn't generic on
anythin except maybe the view type, \XHPRoot).