Lay out generators for comparison conceptually
Generators are the only Hack constructs that can hold state that is not
explicitly typed, while being portable to another scope to resume
execution, possibly with new inputs from `send`. `send` is crucial here,
because these generators (within Route) are being used like simple
objects of a single method to inject "State" at the last second when
rendering or evaluating a write effect. This commit introduces a
synchronous generator as a factory of the asynchronous generators which
generate ViewTrees. This is alternative to cloning asynchronous
generators; behavior like this is necessary to separate the states of
distinct requests, but cloning async generators segfaults at least in
HHVM <= 3.19.2 ([#7802](https://github.com/facebook/hhvm/issues/7802)).

There is also the added benefit that the factory-factory produces a
fresh async generator at render-time, so it doesn't need to yield a
dummy `null` to get its first `State`.